En 5313 refactor trie snapshoting

+++ data/mock/evictionWaitingListMock.go
@@ -11,13 +11,13 @@ import (
 // cacher and db will overwrite the previous values. This structure is not concurrent safe.
 type EvictionWaitingList struct {
 	Cache       map[string][][]byte
-	CacheSize   int
+	CacheSize   uint
 	Db          storage.Persister
 	Marshalizer marshal.Marshalizer
 }
 
 // NewEvictionWaitingList creates a new instance of evictionWaitingList
-func NewEvictionWaitingList(size int, db storage.Persister, marshalizer marshal.Marshalizer) (*EvictionWaitingList, error) {
+func NewEvictionWaitingList(size uint, db storage.Persister, marshalizer marshal.Marshalizer) (*EvictionWaitingList, error) {
 	if size < 1 {
 		return nil, data.ErrInvalidCacheSize
 	}
@@ -38,7 +38,7 @@ func NewEvictionWaitingList(size int, db storage.Persister, marshalizer marshal.
 
 // Put stores the given hashes in the eviction waiting list, in the position given by the root hash
 func (ewl *EvictionWaitingList) Put(rootHash []byte, hashes [][]byte) error {
-	if len(ewl.Cache) < ewl.CacheSize {
+	if uint(len(ewl.Cache)) < ewl.CacheSize {
 		ewl.Cache[string(rootHash)] = hashes
 		return nil
 	}
@@ -88,6 +88,6 @@ func (ewl *EvictionWaitingList) IsInterfaceNil() bool {
 }
 
 // GetSize returns the size of the cache
-func (ewl *EvictionWaitingList) GetSize() int {
+func (ewl *EvictionWaitingList) GetSize() uint {
 	return ewl.CacheSize
 }
